refactor: Repository interface 이동
