+++ app/src/main/java/woowacourse/shopping/datas/ProductDataRepository.kt
@@ -2,6 +2,7 @@ package woowacourse.shopping.datas
 
 import com.shopping.domain.Price
 import com.shopping.domain.Product
+import com.shopping.domain.ProductRepository
 
 object ProductDataRepository : ProductRepository {
     private val baseProducts = listOf(
