+++ app/src/main/java/woowacourse/shopping/productcatalogue/ProductCataloguePresenter.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.productcatalogue
 
 import com.shopping.domain.CartProduct
-import woowacourse.shopping.datas.CartRepository
-import woowacourse.shopping.datas.ProductRepository
-import woowacourse.shopping.datas.RecentRepository
+import com.shopping.domain.CartRepository
+import com.shopping.domain.ProductRepository
+import com.shopping.domain.RecentRepository
 import woowacourse.shopping.mapper.toDomain
 import woowacourse.shopping.mapper.toUIModel
 import woowacourse.shopping.uimodel.ProductUIModel
